Update SmoothMove.cs
Before this commit, idealVector was not rotated by the current rotation of the player, leading to an objective forward, backward, left, etc. Note that Quaternions must come first when multiplying by a vector, and multiplying a quaternion by a vector results in a rotated vector. Anyway, the update now, for instance, allows you to turn and then moves you based on your current rotation such that forward, left, etc are subjective.